requirements.txt
langchain-google-genai
langgraph
python-dotenv
pymongo
pydantic
